api/requirements.txt
fastapi==0.128.8
uvicorn[standard]==0.39.0
python-multipart==0.0.20
onnxruntime>=1.16.0
numpy>=1.24.0,<2
opencv-python-headless>=4.8.0
Pillow>=10.0.0
scipy>=1.11.0
pydantic-settings>=2.1.0
mediapipe>=0.10.14
